implemented forest version of online learner inspired by batch size learning (originally from cicada_learn_block)

+++ cicada/operation/traversal.hpp
@@ -331,7 +331,7 @@ namespace cicada
 	  const int pos_dependent = boost::apply_visitor(__point(), diter->second);
 	  
 	  if (pos_head >= 0 && pos_dependent - 1 >= 0) {
-	    if (pos_dependent - 1 >= yield.size())
+	    if (pos_dependent - 1 >= static_cast<int>(yield.size()))
 	      yield.resize(pos_dependent, -1);
 	    
 	    yield[pos_dependent - 1] = pos_head;
